🔨 Update seed script
